cognito_admin_ui_project/requirements.txt
boto3
python-dotenv
tk
